replace abstact scraper with generic scraper
